requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
pyTelegramBotAPI==4.16.1
python-dotenv==1.0.0
requests==2.31.0
telebot==0.0.5
urllib3==2.2.1
